[feat] Update the useTodoList hook
